+++ R/treatment_effect.R
@@ -16,6 +16,7 @@
 #' - `treatment`: the treatment assignment.
 #' - `variance`: the variance of the treatment effect.
 #' - `jacobian`: the Jacobian matrix.
+#' - `contrast_mat`: contrast summary matrix.
 #'
 #' @export
 treatment_effect <- function(
@@ -35,15 +36,19 @@ treatment_effect.prediction_cf <- function(
   trt_jac <- eff_jacobian(object$estimate[pair[[1]]], object$estimate[pair[[2]]])
   trt_jac_mat <- jac_mat(trt_jac, pair)
   equal_val <- eff_measure(object$estimate[1], object$estimate[1])
+  
 
-  if (is.null(object$variance)) {
+  contrast_var <- trt_jac_mat %*% object$variance %*% t(trt_jac_mat)
+
+  if (is.null(contrast_var)) {
     trt_sd <- rep(NA, length(object))
   } else {
-    trt_sd <- sqrt(diag(object$variance))
+    trt_sd <- sqrt(diag(contrast_var))
   }
+
   z_value <- as.numeric(trt_effect - equal_val) / trt_sd
   p <- 2 * pnorm(abs(z_value), lower.tail = FALSE)
-  coef_mat <- matrix(
+  contrast_mat <- matrix(
     c(
       trt_effect,
       trt_sd,
@@ -52,9 +57,8 @@ treatment_effect.prediction_cf <- function(
     ),
     nrow = length(trt_effect)
   )
-  colnames(coef_mat) <- c("Estimate", "Std.Err", "Z Value", "Pr(>|z|)")
-  row.names(coef_mat) <- sprintf("%s v.s. %s", attr(pair, "levels")[pair[[1]]], attr(pair, "levels")[pair[[2]]])
-
+  colnames(contrast_mat) <- c("Estimate", "Std.Err", "Z Value", "Pr(>|z|)")
+  row.names(contrast_mat) <- sprintf("%s v.s. %s", attr(pair, "levels")[pair[[1]]], attr(pair, "levels")[pair[[2]]])
   structure(
     list(
       estimate = trt_effect,
@@ -64,9 +68,9 @@ treatment_effect.prediction_cf <- function(
       marginal_mean = object,
       fit = object$fit,
       treatment = object$treatment_formula,
-      variance = trt_jac_mat %*% object$variance %*% t(trt_jac_mat),
+      variance = contrast_var,
       jacobian = trt_jac_mat,
-      coef_mat = coef_mat
+      contrast_mat = contrast_mat
     ),
     class = "treatment_effect"
   )
@@ -107,6 +111,15 @@ risk_ratio <- function(object, ...) {
 #' @rdname treatment_effect
 odds_ratio <- function(object, ...) {
   treatment_effect(object, eff_measure = h_odds_ratio, eff_jacobian = h_jac_odds_ratio, ...)
+}
+
+#' @rdname treatment_effect
+log_risk_ratio <- function(object, ...) {
+  treatment_effect(object, eff_measure = h_log_ratio, eff_jacobian = h_jac_log_ratio, ...)
+}
+#' @rdname treatment_effect
+log_odds_ratio <- function(object, ...) {
+  treatment_effect(object, eff_measure = h_log_odds_ratio, eff_jacobian = h_jac_log_odds_ratio, ...)
 }
 
 #' Contrast Functions and Jacobians
@@ -159,9 +172,35 @@ h_odds_ratio <- function(x, y) {
 #' @rdname contrast
 #' @export
 h_jac_odds_ratio <- function(x, y) {
-  assert_numeric(x, lower = 0)
+  assert_numeric(x, lower = 0, upper = 1)
   assert_numeric(y, lower = 0, upper = 1, len = length(x))
   cbind(1 / (1 - x)^2 / y * (1 - y), -x / (1 - x) / y^2)
+}
+
+#' @rdname contrast
+#' @export
+h_log_ratio <- function(x, y) {
+  log(h_ratio(x, y))
+}
+
+#' @rdname contrast
+#' @export
+h_jac_log_ratio <- function(x, y) {
+  cbind(1 / x, -1 / y)
+}
+
+#' @rdname contrast
+#' @export
+h_log_odds_ratio <- function(x, y) {
+  log(h_odds_ratio(x, y))
+}
+
+#' @rdname contrast
+#' @export
+h_jac_log_odds_ratio <- function(x, y) {
+  assert_numeric(x, lower = 0, upper = 1)
+  assert_numeric(y, lower = 0, upper = 1, len = length(x))
+  cbind(1 / x + 1 / (1 - x), -1 / y - 1 / (1 - y))
 }
 
 #' @rdname contrast
@@ -182,6 +221,6 @@ print.treatment_effect <- function(x, level = 0.95, ...) {
   print(x$marginal_mean, signif.legend = FALSE, level = level)
   cat(sprintf("\nContrast     :  %s\n", x$contrast))
   stats::printCoefmat(
-    x$coef_mat
+    x$contrast_mat
   )
 }