+++ NEWS.md
@@ -1,3 +1,11 @@
+# RobinCar2 0.1.2
+
+### New features
+
+* Add `log_odds_ratio` and `log_risk_ratio` function for inference.
+* Add `contrast_mat` to `treatment_effect` object.
+* Add `confint` to obtain confidence interval for `prediction_cf` object.
+
 # RobinCar2 0.1.1
 
 ### New features